* fix compiler warnings
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/branches/LogCaching@9404 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/LogCache/TokenizedStringContainer.cpp
@@ -425,7 +425,7 @@ void CTokenizedStringContainer::AutoCompress()
 	// token pairs. Threashold: log n > n / p
 
 	size_t relation = stringData.size() / ((size_t)pairs.size() + 1);
-	if (stringData.size() < (1 << relation))
+	if (stringData.size() < ((size_t)1 << relation))
 		Compress();
 }
